requirements.txt
opencv-python==4.12.0
numpy==1.26.4
tqdm==4.66.5
requests==2.32.3
beautifulsoup4==4.12.3
pillow==10.3.0